data_filter_azure/requirements.txt
requests
flask
Flask-Bootstrap
git+git://github.com/open-policy-agent/rego-python@master
pytest
azure-storage >= 0.34.3
azure-cosmos 
